// frontend/src/api/types.gen.ts
export type AlbumInfoResponse = {
    /**
     * Album Name
     */
    album_name: string;
    /**
     * Description
     */
    description: string;
    /**
     * Image Urls
     */
    image_urls: Array<string>;
    /**
     * Thumbnail Urls
     */
    thumbnail_urls: Array<string>;
};